fix(all): avoid hydration mismatch for localStorage tools
