+++ app/tools/charcount/ClientOnly.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+const ToolClient = dynamic(() => import("./ToolClient"), { ssr: false });
+
+export default function ClientOnly() {
+  return <ToolClient />;
+}
